+++ lib/subroutine/outputs/output_not_set_error.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Subroutine
+  module Outputs
+    class OutputNotSetError < StandardError
+
+      def initialize(name)
+        super("Expected output '#{name}' to be set upon completion of perform but was not.")
+      end
+
+    end
+  end
+end
