+++ lib/subroutine/outputs/unknown_output_error.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Subroutine
+  module Outputs
+    class UnknownOutputError < StandardError
+
+      def initialize(name)
+        super("Unknown output '#{name}'")
+      end
+
+    end
+  end
+end
