--- lib/subroutine/unknown_output_error.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Subroutine
-  class UnknownOutputError < StandardError
-    def initialize(name)
-      super("Unknown output '#{name}'")
-    end
-  end
-end
