+++ test/subroutine/auth_test.rb
@@ -10,13 +10,13 @@ module Subroutine
     end
 
     def test_it_throws_an_error_if_authorization_is_not_defined
-      assert_raises ::Subroutine::Extensions::Auth::AuthorizationNotDeclaredError do
+      assert_raises ::Subroutine::Auth::AuthorizationNotDeclaredError do
         MissingAuthOp.new
       end
     end
 
     def test_it_throws_an_error_if_require_user_but_none_is_provided
-      assert_raises ::Subroutine::Extensions::Auth::NotAuthorizedError do
+      assert_raises ::Subroutine::Auth::NotAuthorizedError do
         RequireUserOp.submit!
       end
     end
@@ -27,11 +27,12 @@ module Subroutine
 
     def test_it_allows_an_id_to_be_passed
       ::User.expects(:find).with(user.id).returns(user)
-      RequireUserOp.submit! user.id
+      op = RequireUserOp.submit! user.id
+      assert_equal op.current_user, user
     end
 
     def test_it_throws_an_error_if_require_no_user_but_one_is_present
-      assert_raises ::Subroutine::Extensions::Auth::NotAuthorizedError do
+      assert_raises ::Subroutine::Auth::NotAuthorizedError do
         RequireNoUserOp.submit! user
       end
     end
@@ -51,7 +52,7 @@ module Subroutine
     def test_it_runs_custom_authorizations
       CustomAuthorizeOp.submit! user
 
-      assert_raises ::Subroutine::Extensions::Auth::NotAuthorizedError do
+      assert_raises ::Subroutine::Auth::NotAuthorizedError do
         CustomAuthorizeOp.submit! User.new(email_address: "foo@bar.com")
       end
     end
@@ -82,7 +83,7 @@ module Subroutine
     end
 
     def test_it_runs_policies_as_part_of_authorization
-      assert_raises ::Subroutine::Extensions::Auth::NotAuthorizedError do
+      assert_raises ::Subroutine::Auth::NotAuthorizedError do
         PolicyOp.submit! user
       end
 
@@ -101,13 +102,13 @@ module Subroutine
 
       # if: true
       op = IfConditionalPolicyOp.new(user, check_policy: true)
-      assert_raises ::Subroutine::Extensions::Auth::NotAuthorizedError do
+      assert_raises ::Subroutine::Auth::NotAuthorizedError do
         op.submit!
       end
 
       # unless: false
       op = UnlessConditionalPolicyOp.new(user, unless_check_policy: false)
-      assert_raises ::Subroutine::Extensions::Auth::NotAuthorizedError do
+      assert_raises ::Subroutine::Auth::NotAuthorizedError do
         op.submit!
       end
     end
